added profile, search, tabs stacks to stacks folder
