+++ stacks/profileStack.js
@@ -1,0 +1,34 @@
+import Profile from '../screens/Profile';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import { FontAwesome } from '@expo/vector-icons';
+import React from 'react';
+
+const ProfileStack = createNativeStackNavigator();
+const ProfileStackScreen = () => (
+  <ProfileStack.Navigator
+    screenOptions={({ navigation }) => ({
+      headerStyle: {
+        backgroundColor: '#5C6BC0',
+      },
+      headerTintColor: '#fff',
+      headerRight: function headerRight() {
+        return (
+          <FontAwesome
+            name="navicon"
+            size={24}
+            color="white"
+            onPress={() => navigation.toggleDrawer()}
+          />
+        );
+      },
+    })}
+  >
+    <ProfileStack.Screen
+      name="Profile"
+      component={Profile}
+      options={{ headerTitleAlign: 'center' }}
+    />
+  </ProfileStack.Navigator>
+);
+
+export default ProfileStackScreen;
